Implement limited slice operation
This only supports slicing along a single dimension, which is all I
currently need for the models I want to use this library with.

This required fixing a bug in `Tensor::stride`, as it was returning the
stride of the parent of the `dim` dimension, not the stride of the `dim`
dimension itself.