Change image tensor dimensions to match ONNX conventions
Change the dimensions of image tensors from HWC to NCHW, ie. add a batch
dimension and change from channels-last to channels-first. This allows
ONNX models whose tensors conform to the layouts specified in [1] to be
used without having to first permute the dimensions.

It may end up being necessary to change the ordering to optimize
performance, but until the need for that is more certain, it simplifies
things to use the same ordering as ONNX.

[1] https://github.com/onnx/onnx/blob/main/docs/Operators.md