Support in-place execution of operators
Support operators being run in-place on an existing tensor rather than
allocating a new one for the output. This can often be done for element-wise
operators such as ReLU, when the inputs are not going to be needed by other
operators in the graph.

This reduces time in ReLU ops from ~50ms to ~20ms in my test model.