Optimize depthwise Conv2d and ConvTranspose2d with slice indexing
When accessing a range of elements in the last dimension of a tensor, it
is apparently substantially faster to use a slice instead of a
one-dimension index into an UncheckedView. This reduced the time in
Conv2d ops in my test model from ~1100ms to ~800ms and for
ConvTranspose2d ops from ~460ms to ~360ms.