Add additional conv_2d tests using a reference implementation
Using PyTorch to calculate results for all of the conv_2d test cases is
a pain. What we can do is use pre-calculated results for some conv_2d
tests, and also a reference implementation which is optimized for
simplicity. In the existing tests, the pre-calculated results are
compared against the reference implementation as a sanity check.

A specific test has been added for convolutions that are neither
depthwise nor pointwise, since I plan to add a separate code path for
depthwise convolutions compared to the general case soon.