Optimize GEMM with blocking and packing
Iterate over column blocks of the output and then row blocks for each
column. For each column block, pack the corresponding columns of the B
matrix into a buffer in column-major order for more efficient traversal
in the inner loop. Also structure the innermost loop so that the bounds check
on the write to the output matrix can be eliminated.

This reduces the time spent in Conv2d ops from ~780ms to ~700ms on my
test model, where GEMM ops get used in pointwise convolutions.