Change `zero_tensor` and `random_tensor` to take slices rather than Vec
This will avoid an unnecessary allocation if the internal type of the
`shape` field in Tensor is changed in future, eg. to use a smallvec.

I looked into changing `from_data` as well, however that function is
often used in contexts where the `data` and `shape` input Vecs had to be
constructed anyway, so making `from_data` take slices would actually
incur additional copies. For the shape that is not a major issue. For
the data vec it is a problem, since that is often large.