Reduce work on unused space in packing buffers
We only need to pad the columns of the packed B matrix to a multiple of
NR and the rows of the packed A matrix to a multiple of MR, so that we
can feed fixed-sized MRxNR tiles to the kernel. Removing additional
padding in either dimension from the packing buffers should enable some
optimizations to the packing code.