Add a note about a surprising failed experiment
I tried changing these loops to iterate over `0..{used_rows, used_cols}`
since those values are known to be <= H and W. When compiled for x64
this was a minor performance improvement in my tests. When compiled for
wasm it was a more significant regression. Related to this, a set of
benchmarks for matrix multiplication are needed. My testing is still
being done in the context of a complete model run.