Add low-level gemm_slice API that operates directly on slices
This will be useful in some contexts to operate on subsections of tensors or
data that is not held in a tensor. If the package supported creating views of
tensors with different shapes/strides than the original, that would serve as an
alternative to using the low-level API in some of these cases.