Add support for "same" padding to Conv2d operator
Translate the ONNX `auto_pads` attribute to a padding mode enum, which
allows for "same" padding to be specified. ONNX supports two kinds of
"same" padding, to specify which end of each axis gets more padding if
the amount required on each end needs to be different. This library
currently only supports uniform padding on each side.