Add Unsqueeze operator
This initial implementation is based around the older version of the
ONNX Unsqueeze operator where the axes were specified as an attribute
rather than an input, as it is in ONNX v13. This matches the MobileNet
model that I have to hand, but we'll eventually need to move `axes` to
be an input and have the convert-onnx.py script convert attribute values
to graph inputs.