Avoid copying input matrices to Gemm when transposed
Implement the `transpose_a` and `transpose_b` flags for the Gemm operator by
calling `gemm_slice` with swapped row/column strides for the transposed
input(s), instead of copying the input matrices and calling `Tensor::permute` on
them.

For MobileNet fp32 this avoids creating a copy of the ~4.8MB "B" matrix, which
saves ~3ms from the Gemm operation.