Add more flexibility for running operators in-place
Change the type of the mutable input from `&mut Tensor` to `Output`, and the
return type from none to an `Output`. This allow operators to run in-place on
non-float tensors, and also declare general support for running operators
in-place, but bail out and return a different tensor if needed (eg. if
modifying the input is not possible in some circumstances or would be less
efficient than allocating a new tensor.)