Support in-place Add operator
When the inputs to `Add` have exactly the same shape, so no broadcasting
is required, and both are contiguous, run the operator in-place. This saves ~2ms
on MobileNet.

The requirement for the inputs to have a contiguous in-memory layout can
be lifted once `Tensor` provides a way to get a mutable iterator over
elements in logical order, or some alternative way of visiting mutable
references to each element in order.