Document/check for hazard with use of `Tensor::last_dim_slice`
Use of `last_dim_slice` assumes that the last dimension of the tensor is
contiguous. This has been the case in all of my testing to date, but it is now
possible to create tensors where that is no longer the case via
`Tensor::permute`. Some existing uses of `last_dim_slice` could be easily
replaced without significant performance impact. Two other uses require more
work. Add a note about the hazard and an assert to catch it.