Eliminate col2im and temporary output matrix in conv_2d
The col2im step was essentially just reshaping the result of the `kernel_matrix
* im2col_matrix` multiplication from `group_chans x (out_height * out_width)` to
`group_chans, out_height, out_width` and storing the result at the correct place
in the output tensor.

We can eliminate this step by changing the output tensor shape to `[batch,
out_channels, out_height * out_width]` and then performing the matrix
multiplication directly into the appropriate sub-region of this matrix. The
output tensor is then reshaped just before it is returned.

This saves ~5% in Conv2d ops in the Tiny YOLO v2 model (~510ms to ~490ms on my
system).