Eliminate unroll_kernel step in convolution entirely
I realized that the unroll_kernel step is just taking a view of the kernel
tensor and reshaping it to OxIHW. We can do this reshaping implicitly in the
arguments that are passed to `gemm_slice` and thereby eliminate the
`unroll_kernel` helper and `kernel_mat` buffer. This reduces the runtime of Tiny
YOLO v2 by 25-30ms.