Add optional debug names to graph nodes
Having the name of an operator/constant/value from the ONNX model is useful for
debugging issues when running models. Add an optional name to each graph node,
populate it from the `name` field of the corresponding ONNX model node, and
include the name in several debug messages.

In the process the `Node.id` field in the flatbuffers schema has been renamed
to `name`, to match the `name` field of ONNX nodes, which is where the value
comes from.