Separate options for logging timing summary and verbose logging
It turns out that verbose logging has a substantial impact on graph runtime.
MobileNet runs in 60-65ms when logging each op, but ~50ms when not. My text
recognition model runs in ~750ms with each-op logging on but ~650ms with it off.

Therefore separate the options for showing a summary of the graph runtime and
time spent in each op (`timing`) from the option for logging details of each
executed operator.