Add AveragePool2d operator
The ONNX AveragePool operator has very similar attributes to MaxPool,
but they are not the same. The story is similar for the other non-global
pooling operator that ONNX supports, LpPool. This is currently handled
by using a common generic function to provide most of the common
implementation behavior, but separate attribute tables in the schema and
attribute structs in Rust.