Support reading min/max inputs for Clip from attributes or inputs
These inputs to the Clip operator were attributes in ONNX v6 but changed
to inputs in v11. In most models the values for "min"/"max" are still
constants, so we can extract the values from the constant nodes as part
of the model conversion. Models where min/max are actually dynamic are
still not supported.