Add internal support for non-even padding in conv and pooling operators
Support different amounts of padding at the start and ends of axes in
convolution and pooling operators, and use this flexibility to support non-even
padding for `Padding::Same`.

These operators calculate fixed padding and output size for the spatial
dimensions in the same way, from the same inputs. Since there are some subtle
details and notes about the ONNX specs, use a shared implementation.

The model schema and ONNX converter do not yet support non-even padding. This
will be added separately.