Implement a more complete Gather operator
I did not find the explanation in the ONNX specs that clear, but the
examples make it clear that this operator essentially matches the
behavior of `numpy.take` [1], and the docs for that are much more helpful.

In the process a `Tensor::indices` helper was added as a convenient
shorthand for `IndexIterator::from(tensor.shape())`.

[1] https://numpy.org/doc/stable/reference/generated/numpy.take.html#numpy-take