Optimize commutative non in-place binary ops
For commutative ops, it turns out to be faster to copy the larger of the input
operands and then perform the operation in-place on the result, rather than
building up the result item-by-item. This is because the recent optimizations
to in-place binary ops outweigh the costs of doing the initial copy.

This reduces time in Add ops in MobileBERT by ~30-35% (70ms => 45ms).