Add `Tensor::make_contiguous` and fix bug in `Tensor::reshape`
Expose a direct way for `Tensor` users to make it contiguous. This will enable
some operations to use optimized iteration over tensor elements that require a
contiguous layout.

In the process fix a bug in `Tensor::reshape` where the `base` offset was reset
before collecting the elements for the new data buffer, with the result that
`Tensor::elements` would yield the wrong elements.