Optimize `Tensor::elements` for contiguous tensors
Optimize the `elements` implementation for contiguous tensors by making it a
thin wrapper around a slice iterator. This avoids bounds checks each time the
iterator is advanced. A similar optimization was already applied to
broadcasting.

In my text detection image segmentation model, this reduces the time in Concat
ops from ~60ms to ~45ms.