Support negative and out-of-bounds indexes in `Tensor::slice_elements`
In preparation for supporting a more complete version of the Slice
operator, support negative `start` and `end` indexes when slicing a
tensor, as well as out-of-bounds indexes.

Both of these work similarly to slicing in NumPy, since that is where the ONNX
Slice operator semantics come from. Negative indexes count backwards from the
end of a dimension, and out-of-bounds indexes are clamped to the range of valid
indexes, which depends on the step direction.

The `(start, end, step)` tuple slice passed to `slice_elements` has been changed
to a slice of `SliceRange` objects, as this gives the fields useful names and
allows the validation of the step size (must be non-zero) to be done in the
SliceRange constructor.