Support non-float tensors in WASM API inputs and outputs
Change the WASM API to support different types of input and output tensor.
Wasnn's internal generic `Tensor` type is wrapped in a `wasm_api::Tensor` struct
which stores either an int or float tensor, and the `wasm_api::TensorList`
struct stores a list of these tensors.

wasm-bindgen doesn't support APIs that return references to structs, which is a
problem for supporting a `TensorList::get(index)` API, so instead TensorList has
been changed to have a queue-like interface which only supports adding and
removing items. JS code can either use this directly or convert the TensorList
into an array for more convenient usage.