Optimize division by scalar
When the divisor is small compared to the dividend, it makes sense to
convert the operation to multiplication by the reciprocal. A common use
for this is in GELU activations [1] that involve dividing by sqrt(2).

In future we could do this more generally for cases where the second
input to `Div` is a constant.

On EfficientFormerV2 this reduces time in Div ops from ~11ms to ~2ms.

[1] https://paperswithcode.com/method/gelu