Add script to add intermediate tensor values to ONNX model outputs
When debugging differences between ONNX Runtime and Wasnn, it is helpful
to inspect the values of intermediate tensors. Wasnn supports this for
any value in the graph, but ORT only allows model runs to produce
outputs listed in the model's output array. As a workaround, add a
script which adds every operator output node name to the model's list of
outputs.