Fix incorrect broadcasting in binary elementwise ops
Fix an issue where commutative binary elementwise ops (Add, Mul) would
incorrectly broadcast the second argument to the shape of the first if the first
input had a 1-sized dim for which the second input had a non-1 size.

This was due to incorrect logic in `Tensor::can_broadcast`. Fixing this caused a
test for the Expand op, taken from the ONNX spec, to fail. It seems that the
Expand op goes beyond normal NumPy broadcasting. For the moment I have revised
the test to only expect Expand to support our standard broadcasting rules (taken
from NumPy), and added a note about the missing capability.