Optimize binary ops when RHS can be broadcast by cycling
In MobileViT there are many Add operations where the RHS shape matches a
rightmost subset of the LHS. This happens when adding a channel-wise bias to a
channels-last tensor for example. In this case we can simply cycle over RHS
elements. This reduces time in Add ops in MobileViT from ~8 to ~5.5ms.