Add LSTM operator
The initial version supports forwards, reverse and bidirectional LSTMs. It
does not support non-default activations, clipping, custom input layouts or
peepholes.

Since this is much more complex than other operators and involves a lot
of inputs, it is more difficult to programatically construct inputs and
expected results. Therefore two different kinds of tests have been
added:

 1. A test which runs the LSTM operator with random inputs and verifies
    that it successfully produces outputs of the expected shape.

 2. A test which loads inputs, weights and expected outputs generated
    with PyTorch and verifies that this library produces matching
    results.

In the process of writing tests a few new helpers were added to
`Tensor`, and some empty slices in tests in tensor.rs had to be changed
to avoid a conflict with serde_json.