Use the same GEMM kernel for full-sized and non full-sized tiles
For non-full-sized output tiles use a temporary full-sized tile buffer as an
output and then copy the used part back to the real output. In testing with
various models this appears to be just as fast, and it allows us to have a
single kernel implementation in both cases, which will be useful if we introduce
separate kernels in future, for example to take advantage of FMA instructions.