Implement a GEMM kernel that uses AVX + FMA instructions for x64
The auto-vectorized code for the base kernel uses SSE instructions, except when
building for WASM without SIMD, and performs about the same as the SSE `sgemm`
from the matrixmultiply crate. Using AVX + FMA provides a significant
performance boost. Time in matmul ops in MobileViT goes from ~78ms to ~35ms.
The AVX + FMA kernel implementation is very simple, but nevertheless performs
about the same as matrixmultiply's FMA kernel and should be easy to adapt for
other architectures or data types as needed.