Use object API generated by flatc to simplify model conversion
TIL that flatc can generate an object-oriented API for building up data
structures to be serialized, which is much more convenient and concise than the
procedural API. A downside of this API is that it appears that it requires all
fields to also be objects from the object-oriented API, rather than potentially
offsets to already-serialized tables. This limits freedom when mixing and
matching the two API styles, which a caller might want to do to limit memory
usage.

 - Change Makefile to generate object API for Python flatbuffers code
 - Convert operator serialization to use object API. The top-level
   tables are still serialized using the procedural API for now.