Partially support optional inputs in non-trailing positions
ONNX models can have omitted optional inputs and outputs in a non-trailing
position. eg. Given an operator with inputs A, B? C? and outputs X, Y?, Z?,
where "?" denotes optional, it should be possible to omit B and Y, but still
provide C and Z.

This commit adds support for this in the FlatBuffers model format and dataflow
graph. The `Operator::run` API does not yet support this, as it requires some
significant refactoring. As a temporary step, the graph executor will pass an
empty float tensor if an input is omitted.

Omitted input/outputs in non-trailing positions are represented differently
depending on the context:

 1. In the ONNX model, the input or output name is set to an empty string [1]

 2. In convert-onnx.py, entries in an operator node's input/output list are `int|None`s

 3. In the FlatBuffers model, a negative node index is used in
    `OperatorNode.{inputs, outputs}`

 4. In the Rust model graph, entries in an operator node's input/output list
    change from `NodeId` to `Option<NodeId>`.

 5. Finally, in the Operator API, we'll need to change the `&[Input]`
    list to `&[Option<Input>]` or some other struct.

[1] https://github.com/onnx/onnx/blob/main/docs/IR.md#optional-inputs-and-outputs