Rename ops::{activations -> unary_elementwise} and extract Softmax
The most important common characteristic of the ops previously in
ops::activations, from an implementation perspective, is that they take one
input and apply elementwise, not that they are always used as activations in
neural networks.

The exception to this is Softmax, which is much more like a normalization
operator: it retains the shape of its input, but modifies elements based on
statistics computed from many elements.