Convert operator input list from `&[Input]` to InputList
Change the container for operator inputs from a slice of `Input`s to an
`InputList` struct. This solves two problems:

 - It provies a concise syntax in operator implementations for
   extracting arguments and producing appropriate errors if they are
   missing or of the wrong type.

 - It allows making inputs in non-trailing positions optional. This
   could have been done by changing the `inputs` argument to
   `&[Option<Input>]`, but that would have required operators to deal
   with unpacking nested `Options`s. InputList handles this internally.

In the process the two OpError variants for incorrect or unexpected inputs were
combined into one. The current input validation doesn't always make it easy to
know which is the case at the point where the error is generated.