Validate slice length when constructing Matrix
Catch errors when constructing GEMM inputs earlier by validating the slice
length when the `Matrix` view is constructed. The output slice's length is
validated by constructing the implied matrix view for it at the start of the
GEMM operation.

Previously such errors were only caught in the matrix packing functions or GEMM
kernel.