Handle negative endpoint indices in in-place Slice op
Add a `SliceRange::resolve` utility to resolve SliceRange(s) to Range<usize>s
that count forwards or backwards from the start or end of the dimension,
depending on the step, and use it to convert negative slice indices to values
for `Tensor::clip_dim`. An alternative option would have been to make `clip_dim`
support negative indices itself.

In the process revise the docs for SliceRange.