Fix inconsistencies in handling of tensors with adjusted base indexes
The slices returned by `Tensor::{data, data_mut}` are adjusted by the
`Tensor::base` offset, but the offsets returned by `Tensor::offset` were not,
which meant that `tensor.data()[tensor.offset(..)...]` could yield the wrong
elements.

Fix this so that `Tensor::offset` always yields 0 for the first element,
making it valid for use with the slices returned by `data` and `data_mut`.
Also use the `data` method in most places to ensure that the base offset is
taken into account.