Fix incorrect output sizes from Resize op when using `sizes` input
When using `sizes` rather than `scales` to specify the output size, the size was
first converted to a scale as a float. The scale was then converted back to a
size using `floor(in_size * scale)` as specified in the spec. This could result
in a final size which was 1px smaller than the specified size. Fix the issue by
using the opposite approach of using the `sizes` input directly and converting
`scales` to `sizes` using the method from the spec.