Improve handling of invalid input sizes and strides in conv/pool ops
The specs do not specify what should happen if the stride is zero or if the
padded input size is smaller than the kernel size. Here I followed PyTorch's
behavior in producing a runtime error.

This change also fixes an error when running EdgeNeXt in a case where the kernel
size was larger than the un-padded input, but still smaller than the padded
input.