Extract tensor layout methods into TensorLayout trait
Extract the methods for querying tensor shape, strides and offsets from
Tensor into a TensorLayout trait which Tensor, TensorView and TensorViewMut
implement. This makes all of the layout-related methods available for
all three of these structs, at the cost of consumers having to import a
separate trait.

This trait mostly just forwards to the internal `Layout` struct and implementers
just have to provide a method which returns a reference to this.