Change `Tensor::from_data` to take shape as a slice
Tensors no longer store their shape as a Vec internally (instead `Layout` stores
the shape and strides in a `SmallVec<[usize; 8]>`) so it is more efficient and
convenient for this API to take a slice instead.

In the process replace the last remaining non-test uses of the non-associated
`from_data` fn with `Tensor::from_data`.