Support converting attributes to inputs in conversion script
Over time various operator inputs that started out as attributes have been
promoted to dynamic inputs. Amongst currently implemented operators this affects
Clip, Split, Squeeze and Unsqueeze. Handling both forms of providing the input
was previously handled at model conversion time by "down-converting"
constant-valued inputs to attributes. This only worked as long as the input was
actually a constant. This commit replaces that with the opposite approach of
converting the attribute version of an input into a new constant node.