Optimize Concat op for contiguous tensors
Tensor iterators already have a fast path for contiguous tensors, but the Concat
op can do better still by taking advantage of `Vec::extend`'s fast path when
extending with a slice.

This is a further 8-10% improvement in Concat ops in my text detection model
which feeds Concat a mix of contiguous and non-contiguous tensors.