Move Vec + start offset for Tensor into a helper struct
Reduce the number of places that we need to take care of adding the base offset
in Tensor methods by combining the `data` and `base` fields into a helper
struct.

A side effect of this change is that clipping a dimension can now alter both the
start and end of the slice returned by `Tensor::{data, data_mut}`. The slice
length will now always be one more than the highest offset which any valid index
maps to. In other words, the slice "tightly wraps" the used elements within the
data vec. Previously this was only the case for contiguous tensors.