Fix a dumb mistake with lifetimes in `TensorViewMut::{slice, reshape}`
I had used `Self` as a shorthand for `TensorViewMut`, forgetting that `Self`
includes the lifetime generics, so it expanded to `TensorViewMut<'a, T>` instead
of the intended `TensorViewMut<'1, T>` where `'1` is the implicit lifetime of
the `&mut self` receiver. In trying to work around the resulting errors I had
added unnecessary lifetime bounds, which then caused problems when using
`TensorViewMut::{slice, reshape}` elsewhere.