Replace `tensor.last_dim_slice` with `tensor.nd_slice(...).data`
Continue streamlining TensorBase APIs by replacing `last_dim_slice`.

An issue with this change is that `last_dim_slice` used to assert that the data
was contiguous, whereas `NdTensor::data` does not. This should be solved in a
more general way in future by making callers specify whether they expect the
result of data to be contiguous or not.