Simplify `hidden` initialization in LSTM op
The inner loop no longer uses `last_dim_slice` and has been modified to
no longer assume a contiguous `hidden` tensor, although in practice it almost
always be. The new `hidden` => `hidden_seq` copying is slower, but the vast
majority of time in the inner loop is in matmuls, and code to optimize
copying chunks of a tensor can be added later.