Add test for hole border contours
Make the rect used in `test_find_contours_hollow_rect` slightly larger and
thinner so that a hole border is formed, whose shape can be tested.

Hole border points are currently returned in the opposite order (clockwise
instead of counter-clockwise) than outer border points due to the way the
algorithm works. It might make sense to normalize this in future.