Add `line_polygon` helper in page_layout.rs for debugging
For visualizing the boundaries of lines it can be useful to create a polygon
which more closely follows the edges of individual words than using
`min_area_rect` or `convex_hull`. Add a `line_polygon` helper which creates this
polygon. The downside is that the simpler shape returned by `min_area_rect`
looks more attractive.