Improve how `line_polygon` chooses top and bottom edges
When words rects were sufficiently rotated, the previous method could sometimes
choose the rightmost edge as the "top" edge of a rect and then connect the
bottom of that edge to the topmost edge of a subsequent, more horizontal rect,
producing a sharp "cut" in the polygon. Since we rely on the word rects being
ordered left to right, instead always choose the top/bottom edges as the
endpoints of the left/right edges.