Add Polygon type and polygon fill iterator in geometry.rs
Add a Polygon struct which is generic over the storage of vertices, allowing it
to be used as a standalone container or a view of points owned by another shape.

Polygon also has an `iter_points` method which can be used to get the
coordinates of points that fill the polygon, eg. for drawing or using a polygon
as a mask.