Generalize `Tensor::from_data`
Generalize `Tensor::from_data` to `TensorBase::from_data` so that it can be used
to construct views from existing slices. To support convenient constructor of
`Tensor`s from elements in `Vec`s, add `From<Vec<T>>` to the underlying storage
type, `VecWithOffset<T>`, that `Tensor` uses.

A downside of this change is that annotations are required in more places
because the type of the storage argument to `Tensor::from_data` is no longer
concrete.