Implement CTC beam search decoder
Add `ctc` module to crate which contains greedy and beam-search based CTC
sequence decoding algorithms.

The beam search implementation is based on
https://gist.github.com/awni/56369a90d03953e370f3964c826ed4b0, which accompanied
the blog post at https://distill.pub/2017/ctc/, but significantly refactored
for performance. The outputs should be the same as that gist.

I have seen that the beam search output scores can differ slightly from the CTC
decoder in PyTorch's torchaudio package, which is based on [Flashlight
Text](https://github.com/flashlight/text). I haven't finished getting to the
bottom of what causes these differences, but it appears that FT keeps separate
beam states based on whether they end in a blank or not, up until the end of
decoding. This crate's implementation however maintains two scores for each
state at each step, one for ends-in-a-blank and one for does-not-end-in-a-blank.

So far in my testing greedy decoding is much faster and produces identical or
near-identical results, in most cases, to beam search for the OCR model I am
testing with. However having a beam search implementation available is useful
for comparison with other OCR implementations which use this. It also
potentially provides a place to incorporate a language model in future.