Add new GEMM execution API that supports prepacked and virtual inputs
Add support for performing matrix multiplications with inputs that are either
pre-packed or virtual, via a new `GemmExecutor` API.

 - Pre-packed inputs are useful when the same inputs (eg. weights) are
   re-used with multiple matmul operations. This allows the packing cost
   to be paid only once, instead of for each input, at the cost of more
   memory usage.

 - Virtual inputs are useful when data needs to be transformed into a
   large matrix that is solely used in a GEMM operation. This is the
   case with im2col-based convolution for example. Virtual inputs allow
   the large intermediate matrix not to be materialized directly, but
   instead the GEMM operation will request that blocks of it be packed
   on-demand. This avoids the overhead of materializing the full matrix
   and then packing it.

The existing `gemm` function is used for typical/simple GEMM usage with unpacked
input matrices. The GemmExecutor API allows inputs to be provided in different
formats (unpacked, packed, virtual) and also to control other aspects of GEMM
execution.

A downside of these APIs is that because they expose functionality not available
in the standard BLAS interface, code using it becomes more tied to the GEMM
implementation used here.