Combine static and dynamic-dimensional index iterators
Merge `NdIndexIterator` and `IndexIterator` into a single struct which
is generic over the index type, with two type aliases, `NdIndexIterator`
and `DynIndexIterator` for the static and dynamic cases.

This helps with ensuring the logic is consistent between the two, although it
might be necessary to split them again in future for performance reasons.