Make `bounding_rect` function generic and move to imageproc crate
This function is generally useful for getting the bounding rect of a slice of
shapes.

The existing `bounding_box` function can be replaced by creating a polygon from
a point slice and using its `bounding_rect`, although these functions differ
slightly in how they determine the bounding rect. This might need to be
revisited.