Refactor text recognition loop
Make the overall flow easier to follow by splitting out the logic for preparing
a batch of inputs for the text recognition model into a separate function.

In the process I noticed that the padding value for line images was set to 0.5,
which represents a white pixel, rather than -0.5, which represents black. A
black background is used when generating the training data, so use that.