Clamp X coords before calculating character bounding rectangle
When converting positions from the recognition output sequence to X coordinates
in the input image, we may end up with values that are out of bounds if the
position refers to the padding region of the text recognition model input. In
that case the `char_rect` function was returning the min/max coordinates for the
line instead of the character, resulting in unexpectedly tall character boxes.
Fix this by clamping the input X coordinates to `char_rect` so that they are in
range. Also make `char_rect` fail if the input X coords are in range rather than
just substituting the wrong Y coords.