Add `RotatedRect::orient_towards`
For every set of 4 corner points that define a visually unique rotated rect,
there are 4 possible `RotatedRect`s with those corners, but with different "up"
vectors and width/height combinations. `orient_towards` selects one of those 4
rects based on a preferred "up" direction. This is useful for callers that want
corner point order or width/height to assume a particular orientation.