Change `Tensor<T>` storage to an ordinary `Vec<T>`
`Tensor<T>` used a custom `VecWithOffset<T>` storage to enable cheap slice
operations by moving a `base` pointer to set the start of the used region of the
buffer. For the models I'm testing with, very similar performance can be
achieved using a `Vec<T>` by just bulk-moving the new used part of the buffer to
the front and then truncating it.

This change eliminates another difference between `Tensor<T>` and `NdTensor<T>`
and some minor annoyances that come with using a non-standard type for the
storage.