Handle coincident points in the input to `min_area_rect`
When the input to `min_area_rect` contained coincident points, `convex_hull`
produced a hull with edges of of zero length. This lead to infinities in the
parallel / perpendicular axis calculation of `min_area_rect`. Consequently if
all points in the input were coincident, the result was `None` instead of a zero
width/height rect.

 - In `convex_hull`, fix handling of coincident points by completing the TODOs
   about breaking ties when multiple points form the same angle with the X axis,
   by keeping only the points furthest from the point point.

 - In `min_area_rect`, initialize result to the axis-aligned bounding rect of
   the input points. This allows the function to return a result for all
   non-empty inputs.

 - In `min_area_rect`, handle the cases where the hull has zero or one vertices.

 - Add `RotatedRect::{contains, expanded}` for use in `min_area_rect` tests

 - Fix issue with `RotatedRect::from_rect` producing incorrect output due to
   rounding by `Rect::center`