Ensure text lines are only recognized once
While a text line is being recognized, keep a promise in the recognition result
map, so that we don't recognize it a second time if the mouse moves away from
the line and back before the recognition results arrive.

Also add a note about a related issue where recognition output can be incomplete
if a multi-line selection is made, but not every line within the selection is
individually hovered.