Correct handling of `None` outputs from line recognition in WASM API
If recognition produces no output for a text line, represent that with a missing
entry in the `TextLineList` list, instead of skipping that entry and leaving
subsequent results at indices that don't match the input list of lines.

I don't love this API design, since it seems likely that JS consumers will be
prone to forgetting to handle this case.