Correct alignment of text overlay when tab is zoomed
When a tab is zoomed, the coordinates of the document viewport and the captured
tab image no longer align. To fix this, scale coordinates in OCR results by the
inverse of the zoom level, before sending them to the content script.

eg. If the tab zoom is 110%, coordinates of text in the screenshot need to be
scaled by 1 / 1.1 for use in the overlay.