Make grouping of lines into paragraphs work a little better
Previously the algorithm considered a line a candidate for appending to a
paragraph if the line joining the center point of the first line in the
paragraph to the candidate's center point, did not intersect any vertical
separators.

The new algorithm prefers stricter proximity for a paragraph's lines by
requiring that successive lines have some horizontal overlap, and are not
separated by a horizontal separator.