Use ndarray to store constant values in ONNX conversion
Use numpy arrays instead of Python `array.array`s to store constant values in
ONNX conversion. NumPy arrays are more convenient to work with in various ways,
but also conversion of ONNX TensorProtos can be simplified using
`onnx.numpy_helper.to_array`.

In the process add support for converting bool tensors to i32 constants. An
instance of this was found in the GPT-2 ONNX model.