Add `--export-boxes <path>` option to OCR CLI tool
This generates a JSON file with the same structure as the training inputs for an
ML model that is being worked on for layout analysis. This can then be used with
the layout model evaluation scrips in that repo, as a way to compare results
when running the model in PyTorch vs Wasnn.