Correct formulae for calculating "same" padding
The previous formulae incorrectly subtracted the input size only from the kernel
width/height, instead of from the whole of the expression before it. This
resulted in padding values being much larger than they should have been for
certain inputs.

The formulae are now exactly what is in the ONNX docs linked above the function,
except for the use of `saturating_sub` to avoid underflow in certain cases (see
added tests).