Rename {TensorCommon, NdTensorCommon} -> {View, NdView}
Rename the trait that provides read-only operations on tensors to View and
NdView for dynamic and static-dim tensors respectively. I'm not completely
certain about this change, but the idea is to align the name with the `Layout`
trait and better indicate what methods the trait provides. The downside is
that `View` is perhaps too generic.