Support steps when slicing a tensor
Previously steps were supported in the `Slice` operator, but not in `{NdTensor,
Tensor}::slice`.

Change `SliceItem::Range` to use a `SliceRange` to represent a range rather than
the standard `Range*` structs. This allows slice ranges to represent stepped
ranges of an axis. In the `slice_layout` helper, steps are used to adjust the
stride for a dimension.