Add fallible variant of `Tensor::{slice_dyn, slice_mut_dyn}`
In most cases the slice spec is computed by calling code, so an invalid slice
indicates programmer error where a panic is appropriate. In some ONNX operators
however slice indices are calculated from graph operations. In that case we want
the graph run to fail with an `OpError` instead of panicking. There are likely
a number of existing operator implementations where this is not handled
correctly. As a step towards better handling in future, add fallible versions
of the slicing methods in the tensor lib.