Support swapping operands for commutative ops
Allow the graph executor to swap inputs for commutative ops like `Add` and `Mul`
in cases where that enables in-place modification rather than allocating a new
output.

In the DeepLabv3 demo for example, there are several Mul ops where the first
operand is a small `[1, C, 1, 1]` tensor which needs to be broadcast to the
shape of a larger second `[N, C, H, W]` operand. Before this commit the
`Mul::run_in_place` impl would allocate a new output when handling these ops,
because only the first operand is mutable.  After this commit, the graph
executor will pick the largest operand, by number of elements, to use as the
output. This reduced time in `Mul` ops in DeepLabv3 from ~4ms to ~1ms.