Add AVX-512 GEMM kernel
This is a copy and paste of the FMA kernel, but with references to 256-bit
regs/instructions changed to 512, and the `NR` constant doubled to match the
register-width doubling. On my Ice Lake system, AVX-512 is a modest speedup
over AVX 2. AVX-512 is also enabled as that gives us more registers.

There is an issue with AVX-512 feature detection on macOS due to how the Darwin
kernel handles enabling AVX 512. See https://github.com/golang/go/issues/43089
and https://github.com/numpy/numpy/pull/19362.