Optimize broadcasting of rightmost dims in binary ops
Optimize broadcasting when the dimensions to be broadcast are on the right. eg.
`[N, C, H, W]` * `[1, C, 1, 1]`. This pattern came up for Add and Mul ops in the
DeepLabv3 and Detr examples. There was an existing fast path for broadcasting
leading dimensions only, but not trailing dims. Extend and optimize the fast
paths for broadcasting to handle the existing cases + this one in a unified
way.