Decrease default approx-equality tolerance from 1e-4 to 1e-5
This tighter tolerance is appropriate for most tests. There are a number of
existing tests where the expected values are specified to a low level of
precision and fail after decreasing the default. To enable migration, allow
custom tolerance values to be specified via `expect_equal_with_tolerance` and
modify these tests to use the old 1e-4 value.