Use a faster Erf implementation
Erf frequently shows up as a slow operator in models that use the Gelu
activation.

Add a faster Erf implementation which matches what PyTorch uses for its
vectorized implementation in aten [1]. This is a well known approximation that
is documented on Wikpedia. It runs approximately ~3x faster than `libm::erff`
and importantly, is easier to vectorize in future. The downside is lower
accuracy, per notes in the tests.

[1] https://github.com/pytorch/pytorch/blob/871e27a61c0f0bf49dab51f68534c49c95ee0ee3/aten/src/ATen/cpu/vec/vec256/vec256_float.h#L160