Optimize `ReduceMean` with more efficient summing of `[f32]`
The default codegen for `slice.iter().sum::<f32>()` is not optimal for long
slices. To get the best results will probably require explicit SIMD usage, but
we can get a significant improvement by using a version that hints to the
compiler that it should use autovectorization.

This change results in about a ~2x reduction in time in ReduceMean ops in
convnext (~11ms => ~5ms).